Refresh zend_mm shadow key on fork
The memory manager is cleaned up after each request by calling
shutdown_memory_manager(). At the same time, this prepares the manager for the
next request, and the shadow key is refreshed.

In forking SAPIs the first request of every child process inherits the memory
manager of the parent process, including the shadow key. As a result, a leak
of the shadow key during the first request of one process gives away the
shadow key used during the first request of other processes. This makes the key
refresh mechanism less useful.

Here I ensure that we refresh the shadow key after a fork. The memory manager is
not empty at this point (we perform allocations after
shutdown_memory_manager()), so we have to recompute any shadow pointers with the
new key.

We assume that the parent process had only one PHP thread running, as anything
else would be unsafe. If a SAPI forks a process running more than one PHP
thread, it is its responsibility to call refresh_memory_manager() in each PHP
thread of the child process.

Closes GH-16765