ext/curl: Fix `CURLOPT_PREREQFUNCTION` to not continue on error conditions
This is a follow-up to GH-13255.

When the `CURLOPT_PREREQFUNCTION` callback does not return any value
or return an invalid value, we throw a `ValueError` or a `TypeError`
exception. However, it does not prevent the Curl request from going
forward because our default return value is `CURL_PREREQFUNC_OK`.

This changes the default value to `CURL_PREREQFUNC_ABORT`.