ext/iconv/tests/bug52211.phpt: skip for non-GNU iconv implementations
This test ensures that iconv() will not return part of the string if
it encounters an error, but the particular error it wants to cause is
not portable. When the input sequence is not representable in the
output charset, POSIX says that "iconv() shall perform an
implementation-defined conversion on the character" rather than
fail. As far as I know, the only two implementations that go against
this recommendation (and return the error that this test is expecting)
are GNU libiconv and glibc.

This commit adds a SKIPIF check for the two GNU implementations to
ensure that an error is actually encountered during this test.