ext/iconv/tests/bug76249.phpt: drop //IGNORE suffix
This test performs a conversion from UCS-2 to UTF-8 using iconv and
stream filters, with the suffix "//IGNORE" being added to the target
charset. This magic //IGNORE string was recently standardized in POSIX
2024, but it is not yet portable: musl for example will simply fail
when it encounters //IGNORE in what is supposed to be a charset name.

Fortunately, we do not need to think too hard about the general
problem here: the input is "abc", which has no un-translatable
sequences. We can simply drop the "//IGNORE".