ext/iconv/tests/bug48147.phpt: skip for non-GNU iconv implementations
This test meaningfully uses the //IGNORE charset suffix that was was a
GNU/Solaris extension but is now standardized in POSIX 2024. The way
//IGNORE is used, however, is non-standard: POSIX says that //IGNORE
will cause untranslatable sequences to be skipped, but this test is
expecting it to skip input sequences that are invalid rather than
unexpressible in the target charset. That behavior is specific to the
two GNU implementations (and was always non-conforming...), so we add
a SKIPIF block to ensure that one of the GNU implementations is used.