Move iconv const check into autoconf
Some systems (older NetBSD pre-10, Solaris) may have a non-standard
const parameter with iconv. Instead of hardcoding the OS check, move
this to a feature check performed by autoconf.

autoconf doesn't have a nicer way of checking this (well, except for
AM_ICONV, which is part of gettext and we're presumably not using it),
so we have to repeat the function signature and check for it with a
mismatched signature.