ext/iconv/tests/bug48147.phpt: rename and repurpose
This test ensures that iconv() with //IGNORE will return part of the
translated string when it is given invalid inputs. POSIX and the PHP
documentation however agree that an error should be returned in that
case: the test is making sure that we get the wrong answer.

PHP's iconv has recently been updated to agree with the standard and
docs. As a result, this test fails, but that's OK -- it should. The
history of iconv in PHP is convoluted, so rather than just update the
expected output here, I think it's best to rename the test and sever
its ties to bug 48147. Referencing that bug at this point can only
cause confusion.