CI: FreeBSD on VM - set `copyback: false`
The `vmactions/freebsd-vm` GitHub action rsyncs the work dir to to the
VM. This adds a lot of log output due to `rsync -v` usage.

Once the tests are compelte, the action copies the files _back_ by
running `rsync` in reverse. However, we do not need these files back
because we do not run any other steps that need access to the post-test
files.

Setting `copyback: false` disables this, and cuts the log size by about
5,000 lines.

Closes Closes GH-16916.