test: persistent curl share handles
I opted to use the existing Caddy testing infrastructure since it
supports keepalives, whereas it seems the PHP development server does
not.

Alternatively I could write just enough of a socket listener to confirm
that there is only ever a single connection coming from curl, but I
believe it is safe to rely on connect_time being zero when a connection
is reused.