allow urlencode(int)
in strict mode it's kindof annoying to have to cast-to-string when you have a string|int variable you need to urlencode.
There is no ambiguity as to how to urlencode integers, so it shouldn't be a problem.
"TypeError: urlencode(): Argument #1 ($string) must be of type string, int given"
Has happened to me enough times now that I bothered to make PR to allow it.

Fwiw the integer path could be micro-optimized to bypass php_raw_url_encode() entirely, at least when the integer is not negative,
but for brewity, I didn't bother micro-optimizing it.