Fix mbstring/libmbfl/config.h
This moves libmbfl configuration to mbstring extension level and syncs
the include style used as of PHP-8.4 to be able to do simultaneous
in-source and out-of-source builds. The HAVE_CONFIG_H file is when doing
phpize build using Autotools.

Checks for strcasecmp and strings.h are also moved to mbstring extension
only.