Fix mbstring/libmbfl/config.h
This moves libmbfl configuration to mbstring extension level and syncs
the include style used as of PHP-8.4 to be able to do simultaneous
in-source and out-of-source builds. The HAVE_CONFIG_H file is when doing
phpize build using Autotools.

Checks for strcasecmp and strings.h are also moved to mbstring extension
only.

The libmbfl/config.h header is now also NOT installed anymore as it is a
private header used only in C files as part of the build interface
(during build phase only).

The libmbfl/config.h is included with angled style so the include flags
take precedence and simultaneous php-src building in-source and
out-of-source can be performed.