Actually use cwd_skip
We should strip the CWD whenever we call `php_glob_stream_path_split()`,
but that would require to store the value in `glob_s_t` what we cannot
do for stable PHP versions since it would be an ABI break.  Thus we
traverse `gl_pathv`, and strip the prepended CWD there.